tools/requirements.txt
# Main Application & ADK
google-adk>=1.0.0
fastapi
uvicorn[standard]
python-dotenv>=1.0.0

# Google Cloud & AI
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
firebase-admin

# Data & Utilities
redis>=5.0.0
pymongo>=4.0.0
requests>=2.31.0
pydantic
reportlab